collapse to_struct intoo serialization
